refactor: split event pagination controls
