+++ src/Pages/Events/EventsPage.js
@@ -2,20 +2,19 @@ import { useState, useEffect, useMemo, useRef } from "react";
 import mockEvents from "./eventsMockData.json";
 import EventHero from "./EventHero";
 import EventCard from "./EventCard";
-import { ChevronLeft, ChevronRight, Grid, List } from "lucide-react";
+import { Grid, List } from "lucide-react";
 import FeedbackButton from "../../components/FeedbackButton";
 import EventCTA from "./EventCTA";
 import Fuse from "fuse.js";
 import StyledDropdown from "../../components/StyledDropdown";
 import { EventCardSkeleton } from "../../components/common/SkeletonLoaders";
+import PaginationControls from "./PaginationControls";
 import {
   DEFAULT_EVENTS_PER_PAGE,
-  EVENTS_PER_PAGE_OPTIONS,
   clampPage,
   filterEventsByType,
   getPaginatedEvents,
   getTotalPages,
-  getVisiblePaginationPages,
   sortEventsByDate,
 } from "./eventPaginationUtils";
 
@@ -59,125 +58,6 @@ const renderCardSection = (isLoading, eventsToShow, viewMode, filterType) => {
   );
 };
 
-const PaginationControls = ({
-  currentPage,
-  eventsPerPage,
-  totalEvents,
-  totalPages,
-  onPageChange,
-  onPageSizeChange,
-}) => {
-  if (totalEvents === 0) {
-    return null;
-  }
-
-  const startEvent = (currentPage - 1) * eventsPerPage + 1;
-  const endEvent = Math.min(currentPage * eventsPerPage, totalEvents);
-  const { firstVisiblePage, lastVisiblePage, pages: visiblePages } =
-    getVisiblePaginationPages(currentPage, totalPages);
-
-  return (
-    <div className="mt-10 flex flex-col gap-4 border-t border-gray-200 pt-6 dark:border-gray-700 sm:flex-row sm:items-center sm:justify-between">
-      <p className="text-sm text-gray-600 dark:text-gray-400">
-        Showing {startEvent}-{endEvent} of {totalEvents} events
-      </p>
-
-      <div className="flex flex-col gap-3 sm:flex-row sm:items-center">
-        <label
-          htmlFor="events-per-page"
-          className="flex items-center gap-2 text-sm text-gray-600 dark:text-gray-400"
-        >
-          Per page
-          <select
-            id="events-per-page"
-            value={eventsPerPage}
-            onChange={(event) => onPageSizeChange(Number(event.target.value))}
-            className="rounded-lg border border-gray-200 bg-white px-3 py-2 text-sm text-gray-700 shadow-sm focus:border-black focus:outline-none focus:ring-2 focus:ring-black/10 dark:border-gray-700 dark:bg-gray-800 dark:text-gray-200"
-          >
-            {EVENTS_PER_PAGE_OPTIONS.map((option) => (
-              <option key={option} value={option}>
-                {option}
-              </option>
-            ))}
-          </select>
-        </label>
-
-        <nav className="flex items-center gap-2" aria-label="Event pagination">
-          <button
-            type="button"
-            onClick={() => onPageChange(currentPage - 1)}
-            disabled={currentPage === 1}
-            className="inline-flex h-10 w-10 items-center justify-center rounded-lg border border-gray-200 bg-white text-gray-700 transition hover:bg-gray-50 disabled:cursor-not-allowed disabled:opacity-40 dark:border-gray-700 dark:bg-gray-800 dark:text-gray-200 dark:hover:bg-gray-700"
-            aria-label="Previous page"
-          >
-            <ChevronLeft size={18} />
-          </button>
-
-          {firstVisiblePage > 1 && (
-            <>
-              <button
-                type="button"
-                onClick={() => onPageChange(1)}
-                className="h-10 min-w-10 rounded-lg border border-gray-200 bg-white px-3 text-sm font-medium text-gray-700 transition hover:bg-gray-50 dark:border-gray-700 dark:bg-gray-800 dark:text-gray-200 dark:hover:bg-gray-700"
-              >
-                1
-              </button>
-              {firstVisiblePage > 2 && (
-                <span className="px-1 text-sm text-gray-500 dark:text-gray-400">
-                  ...
-                </span>
-              )}
-            </>
-          )}
-
-          {visiblePages.map((page) => (
-            <button
-              type="button"
-              key={page}
-              onClick={() => onPageChange(page)}
-              className={`h-10 min-w-10 rounded-lg px-3 text-sm font-medium transition ${
-                page === currentPage
-                  ? "bg-black text-white"
-                  : "border border-gray-200 bg-white text-gray-700 hover:bg-gray-50 dark:border-gray-700 dark:bg-gray-800 dark:text-gray-200 dark:hover:bg-gray-700"
-              }`}
-              aria-current={page === currentPage ? "page" : undefined}
-            >
-              {page}
-            </button>
-          ))}
-
-          {lastVisiblePage < totalPages && (
-            <>
-              {lastVisiblePage < totalPages - 1 && (
-                <span className="px-1 text-sm text-gray-500 dark:text-gray-400">
-                  ...
-                </span>
-              )}
-              <button
-                type="button"
-                onClick={() => onPageChange(totalPages)}
-                className="h-10 min-w-10 rounded-lg border border-gray-200 bg-white px-3 text-sm font-medium text-gray-700 transition hover:bg-gray-50 dark:border-gray-700 dark:bg-gray-800 dark:text-gray-200 dark:hover:bg-gray-700"
-              >
-                {totalPages}
-              </button>
-            </>
-          )}
-
-          <button
-            type="button"
-            onClick={() => onPageChange(currentPage + 1)}
-            disabled={currentPage === totalPages}
-            className="inline-flex h-10 w-10 items-center justify-center rounded-lg border border-gray-200 bg-white text-gray-700 transition hover:bg-gray-50 disabled:cursor-not-allowed disabled:opacity-40 dark:border-gray-700 dark:bg-gray-800 dark:text-gray-200 dark:hover:bg-gray-700"
-            aria-label="Next page"
-          >
-            <ChevronRight size={18} />
-          </button>
-        </nav>
-      </div>
-    </div>
-  );
-};
-
 const EventsPage = () => {
   const [events, setEvents] = useState([]);
   const [filterType, setFilterType] = useState("all");
@@ -207,7 +87,10 @@ const EventsPage = () => {
       ? fuse.search(searchQuery).map((res) => res.item)
       : events;
 
-    return sortEventsByDate(filterEventsByType(searchResults, filterType), sortType);
+    return sortEventsByDate(
+      filterEventsByType(searchResults, filterType),
+      sortType
+    );
   }, [events, filterType, searchQuery, sortType]);
 
   const totalPages = getTotalPages(filteredEvents.length, eventsPerPage);
@@ -267,7 +150,7 @@ const EventsPage = () => {
               { key: "past", label: "Past" },
               { key: "conference", label: "Conferences" },
               { key: "workshop", label: "Workshops" },
-            ].map((filter, index) => (
+            ].map((filter) => (
               <button
                 key={filter.key}
                 onClick={() => setFilterType(filter.key)}
